Another four updated DynamoDB examples.
