--- dotnetv3/dynamodb/FromSQL/GetOrdersForProductTest/GetOrdersForProductTest.csproj
@@ -1,23 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-
-    <IsPackable>false</IsPackable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="AWSSDK.Core" Version="3.5.1.17" />
-    <PackageReference Include="AWSSDK.DynamoDBv2" Version="3.5.0.19" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.5.0" />
-    <PackageReference Include="moq" Version="4.14.5" />
-    <PackageReference Include="xunit" Version="2.4.0" />
-    <PackageReference Include="xunit.runner.visualstudio" Version="2.4.0" />
-    <PackageReference Include="coverlet.collector" Version="1.2.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\GetOrdersForProduct\GetOrdersForProduct.csproj" />
-  </ItemGroup>
-
-</Project>
