--- dotnetv3/dynamodb/FromSQL/ListTables/ListTables.csproj
@@ -1,20 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-	
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="AWSSDK.Core" Version="3.5.1.3" />
-    <PackageReference Include="AWSSDK.DynamoDBv2" Version="3.5.0.5" />
-    <PackageReference Include="System.Configuration.ConfigurationManager" Version="4.7.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Update="app.config">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
-
-</Project>
