+++ dotnetv3/dynamodb/FromSQL/ListTablesExample/app.config
